% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_abs.R
\name{int_abs}
\alias{int_abs}
\title{Create an absorbance measurement from one file of transmittance and one file of reflectance}
\usage{
int_abs(trans_loc, ref_loc, averaging = 1)
}
\arguments{
\item{trans_loc}{location of a file with transmittance data}

\item{ref_loc}{location of a file with reflectance data}
}
\description{
this function accepts a single transmission dataset, a single reflection dataset,
and returns the absorbance calculated as a=1-t-r
}
