% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_baseline_all.R
\name{int_baseline_all}
\alias{int_baseline_all}
\title{Fully work up a complete data set of integrating sphere data}
\usage{
int_baseline_all(
  locationBaseline,
  locationReflectance,
  locationTransmittance,
  Averaging = 1,
  writeLoc = NULL,
  writePrefix = "",
  licordat = T,
  useSharkeySpec = F
)
}
\arguments{
\item{Averaging}{a variable that reflects how many wavelengths will be averaged together. Default: 1, so 1 point per wavelength}

\item{writeLoc}{If provided, will write out the collected data to the chosen directory. Provides csvs for reflectance, transmittance, baseline, and compiled.}

\item{writePrefix}{Standard text added to the front of filenames for written out files.}

\item{licordat}{Show or hide the Li-6800 LED spectra}
}
\description{
A fully-featured function that accepts a series of folders to automatically work up.
Pass in the FOLDER location of your baseline (no plant) reflectance, plant reflectance
and plant transmittance.
}
