% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_graph.R
\name{int_graph}
\alias{int_graph}
\title{Turn a integrating sphere dataset into a nice graph with licor dataset}
\usage{
int_graph(dataset, licordat = TRUE)
}
\arguments{
\item{dataset}{A dataset with Wavelength, Transmittance, Reflectance, and Absorbance data}

\item{licordat}{show or hide the li-6800 LED spectra}
}
\description{
A very narrow function because it is written specially for my own computer.
accepts a compiled dataset with specifically 4 columns, in order.
designed to be called from other functions in this file...
returns a ggplot object.
Requires reshape2 library.
displays a graph with the transmittance reflectance, absorbance and the licor LED spectra
and also adds annotations dictating the average absorbance over the half-bandwidths of the blue and red LEDs.
}
