% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_read_all.R
\name{int_read_all}
\alias{int_read_all}
\title{Read all data in the folder, parse it based on type, and average them}
\arguments{
\item{tpattern}{text check for transmission data}

\item{rpattern}{text check for reflection data}
}
\description{
Given a FOLDER location, returns the average of all the transmissions and reflectances
in that folder.  This is used to average a significant number of scans of the same
sample
could easily be rewritten to not average them...
}
