% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_read_one.R
\name{int_read_one}
\alias{int_read_one}
\title{Reads one file and formats it for use with other functions}
\arguments{
\item{location}{location of an integrating sphere data file}
}
\description{
simple function that interprets the raw data file
"Averaging" means you get one data point per # wavelengths
so, with default, you get one data point per wavelength, usually the average of 2-3
data points
}
