% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_reconstitute.R
\name{int_reconstitute}
\alias{int_reconstitute}
\title{Restore data written by int_baseline_all}
\description{
This function uses the written data from int_baseline_all to restore the analysis but
without requiring that all the files be read in again.
}
